requirements.txt
pandas
streamlit
paho-mqtt
pillow
numpy
matplotlib
plotly